Update imports for renamed financial symbol processor
