+++ CorpusBuilderApp/shared_tools/ui_wrappers/processors/financial_symbol_processor_wrapper.py
@@ -1,7 +1,7 @@
 # File: shared_tools/ui_wrappers/processors/financial_symbol_processor_wrapper.py
 
 from PySide6.QtCore import Signal as pyqtSignal, QThread
-from shared_tools.processors.finacial_symbol_processor import FinancialSymbolProcessor
+from shared_tools.processors.financial_symbol_processor import FinancialSymbolProcessor
 from shared_tools.ui_wrappers.base_wrapper import BaseWrapper
 from shared_tools.processors.mixins.processor_wrapper_mixin import ProcessorWrapperMixin
